fix(cuda): smoke_fill routed the fill value through float32 before the cast
static_cast<scalar_t>(static_cast<float>(value)) narrowed to float32 and back
regardless of scalar_t, so a float64 tensor silently got a float32-rounded
value instead of the double it was given. static_cast<scalar_t>(value) directly
narrows correctly for every dispatched type, including double.

+++ csrc/smoke.cu
@@ -54,7 +54,7 @@ void smoke_fill(at::Tensor tensor, double value) {
   AT_DISPATCH_FLOATING_TYPES_AND2(at::ScalarType::Half, at::ScalarType::BFloat16,
                                   tensor.scalar_type(), "smoke_fill", [&] {
     smoke_fill_kernel<scalar_t><<<blocks, kThreads, 0, stream>>>(
-        tensor.data_ptr<scalar_t>(), numel, static_cast<scalar_t>(static_cast<float>(value)));
+        tensor.data_ptr<scalar_t>(), numel, static_cast<scalar_t>(value));
   });
   C10_CUDA_KERNEL_LAUNCH_CHECK();
 }
